requirements.txt
fastapi >= 0.115.14
uvicorn >= 0.34.3
python-dotenv
